refactor prisma commands into seperate folders
