+++ src/app/dashboard/_components/_video-files/video-context-menu.tsx
@@ -1,5 +1,4 @@
 import { ContextMenuContent, ContextMenuItem, ContextMenuSeparator, ContextMenuSub, ContextMenuSubContent, ContextMenuSubTrigger, } from '@/components/ui/context-menu'
-import { ContextMenu } from '@/components/ui/context-menu'
 import React from 'react'
 import { cn } from '@/lib/utils'
 import { ChevronDown, ChevronUp, Eye, EyeOff, FolderInput } from 'lucide-react'
