+++ src/app/dashboard/_components/_video-files/video-file.tsx
@@ -9,10 +9,10 @@ import { FileEntry } from "@tauri-apps/api/fs";
 import type { User, Video } from "@prisma/client";
 import {
   closeDatabase,
-  updateVideoWatched,
-} from "../../../../../lib/prisma-commands";
+} from "../../../../../lib/prisma-commands/misc-cmds";
 import { invoke } from "@tauri-apps/api/tauri";
 import VideoContextMenu from "./video-context-menu";
+import { updateVideoWatched } from "../../../../../lib/prisma-commands/folders/folder-cmds";
 
 export default function VideoFile({
   userSettings,
@@ -109,14 +109,14 @@ export default function VideoFile({
               className={cn(
                 "text-base flex flex-row items-start justify-center gap-1 font-medium select-none text-center",
                 file.name &&
-                  file.name?.length > 20 &&
-                  "overflow-hidden whitespace-nowrap",
+                file.name?.length > 20 &&
+                "overflow-hidden whitespace-nowrap",
               )}
               key={"current-video-file-name-motion-div" + file.name + index}
               whileHover={
                 userSettings?.animations === "On" &&
-                file.name &&
-                file.name?.length > 65
+                  file.name &&
+                  file.name?.length > 65
                   ? { width: "-100%" }
                   : undefined
               }
@@ -126,8 +126,8 @@ export default function VideoFile({
                 className={cn(
                   "h-auto w-3",
                   file.name &&
-                    file.name?.length > 100 &&
-                    "items-start justify-center gap-1 p-0",
+                  file.name?.length > 100 &&
+                  "items-start justify-center gap-1 p-0",
                   userSettings?.fontSize === "Medium" && "h-auto w-3.5",
                   userSettings?.fontSize === "Large" && "h-auto w-4",
                   userSettings?.fontSize === "XLarge" && "h-auto w-5",
@@ -212,7 +212,7 @@ export default function VideoFile({
                         "h-auto w-5 mr-0.5 ",
                         userSettings?.fontSize === "Medium" && "h-auto w-5",
                         userSettings?.fontSize === "Large" &&
-                          "h-auto w-[1.3.5rem]",
+                        "h-auto w-[1.3.5rem]",
                         userSettings?.fontSize === "XLarge" && "h-auto w-7",
                       )}
                     />
@@ -260,11 +260,11 @@ export default function VideoFile({
             </motion.div>
           </motion.li>
           {index === files.length - 1 && (
-            <div className="relative h-1.5 w-auto flex flex-row items-center justify-center gap-1 rounded-b-full"
+            <div className="relative flex h-1.5 w-auto flex-row items-center justify-center gap-1 rounded-b-full"
 
-							style={{ backgroundColor: `${currentFolderColor}`}}
+              style={{ backgroundColor: `${currentFolderColor}` }}
 
-						/>
+            />
           )}
           {/* VideoContextMenu was moved out of the motion.li, if something is wrong it might be this */}
