+++ src/app/login/page.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import React, { useState, createRef, useEffect, use } from 'react';
-import { closeDatabase, getCurrentUserGlobal, getUsers, setCurrentUserGlobal } from '../../../lib/prisma-commands';
+import { closeDatabase } from '../../../lib/prisma-commands/misc-cmds';
 import type { User } from "@prisma/client";
 import { useRouter } from 'next/navigation';
 import { UserAvatar } from '../profiles/_components/user-avatar';
@@ -9,6 +9,8 @@ import { Button } from '@/components/ui/button';
 import { Check } from 'lucide-react';
 import { AnimatePresence, motion } from 'framer-motion';
 import { cn } from '@/lib/utils';
+import { getCurrentUserGlobal, setCurrentUserGlobal } from '../../../lib/prisma-commands/global/global-cmds';
+import { getUsers } from '../../../lib/prisma-commands/user/user-cmds';
 
 export default function Home() {
     let router = useRouter();
